+++ UnityProject/Assets/Resources/Prefabs/Objects/Wallmounts/DoorSwitch.prefab.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 1325f48af15d4163ab952de2fcb5a184
+timeCreated: 1480237788
+licenseType: Pro
+NativeFormatImporter:
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
