--- UnityProject/Assets/Scripts/Doors/ShutterController.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: c3e79b1b3e974214482328792b14c76d
-timeCreated: 1480931764
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
