--- UnityProject/Assets/Scripts/Doors/WinDoorAnimator.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: c2cf30704fd754096ad1a718fc9b2782
-timeCreated: 1506407108
-licenseType: Pro
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
