--- UnityProject/Assets/Scripts/Switches/ShutterSwitch.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: aef020156f74075429ef9e6479907b71
-timeCreated: 1480942162
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
